+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Scenarios can now use regular expressions for the delivery configuration, for example `.*` to apply to all
 - `enquire_deliveries_by_scenario` action now lists which deliveries are enabled by the scenario, disabled by it, and all the scenarios to which it applies overrides
 - `delivery` section now has the same flexibility as on an action call - it can be a mapping, list or single string, the latter two are all that is needed to simply enable deliveries
-#### Media
+#### Cameras
+- Camera entity's built in device tracker will now be used, no need for separate device tracker where this supported
 - Better diagnostics for unavailable cameras
 #### Notification Archive
 - Further improvements to `Notification` archive object to make it easier to debug